requirements.txt
# Core FastAPI and web framework
fastapi>=0.104.0
uvicorn>=0.24.0

# Document processing
PyMuPDF>=1.23.0
pdfplumber>=0.7.6
python-docx>=0.8.11
beautifulsoup4>=4.12.0

# AI/ML libraries
openai>=1.0.0
google-generativeai>=0.3.0
groq>=0.4.0
sentence-transformers>=2.2.0

# ArXiv integration
arxiv>=2.0.0

# Frontend
streamlit>=1.20.0

# Utilities
python-dotenv>=0.21.0
numpy>=1.22.0
requests>=2.28.0
tqdm>=4.64.0